requirements.txt
# Core dependencies
torch>=2.5.0,<3.0.0
torchvision>=0.20.0,<1.0.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0,<5.0.0
Pillow>=10.0.0,<11.0.0

# Image restoration libraries
basicsr>=1.4.2,<1.5.0
realesrgan>=0.3.0,<0.4.0
facexlib>=0.3.0,<0.4.0

# Utilities
tqdm>=4.65.0,<5.0.0
pyyaml>=6.0,<7.0
scikit-image>=0.21.0,<0.23.0

# Development dependencies
pytest>=7.4.0,<8.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
